Load and provide the checkout token on demand
