+++ example/src/main/kotlin/com/example/afterpay/data/CheckoutResponse.kt
@@ -1,5 +1,6 @@
 package com.example.afterpay.data
 
 data class CheckoutResponse(
-    val url: String
+    val url: String,
+    val token: String
 )
